+++ codegen/src/rust/names.rs
@@ -74,9 +74,19 @@ pub fn service_event(service: &str) -> String {
     format!("{service}Event")
 }
 
+/// Returns the name of a service's event handler trait.
+pub fn service_event_handler(service: &str) -> String {
+    format!("{service}EventHandler")
+}
+
 /// Returns the name of a service's call type.
 pub fn service_call(service: &str) -> String {
     format!("{service}Call")
+}
+
+/// Returns the name of a service's call handler trait.
+pub fn service_call_handler(service: &str) -> String {
+    format!("{service}CallHandler")
 }
 
 /// Returns the name of a service's introspection type.